Factor common test model setup into base class
